chore: divide test
